# Don't make registerGenerators error unless needed.