This PR avoids allocating an error inside registerGenerators unless it's used. Before, we were allocating it in all cases.

```
$ go test -run=$^ -bench=. -count=5 | tee /tmp/baseline.txt
$ go test -run=$^ -bench=. -count=5 | tee /tmp/experiment.txt
$ benchstat /tmp/{baseline,experiment}.txt
name                           old time/op    new time/op    delta
NewWorkload/NoCallsNoGen-128      482ns ± 1%     486ns ± 1%     ~     (p=0.111 n=5+5)
NewWorkload/NoCalls-128          2.25µs ± 2%    1.67µs ± 2%  -25.98%  (p=0.008 n=5+5)
NewWorkload/OneCall-128          2.32µs ± 1%    1.74µs ± 1%  -25.00%  (p=0.008 n=5+5)
NewSimulator/NoCallsNoGen-128    22.4µs ± 2%    22.1µs ± 4%     ~     (p=0.421 n=5+5)
NewSimulator/NoCalls-128         25.0µs ± 3%    23.5µs ± 1%   -5.92%  (p=0.008 n=5+5)
NewSimulator/OneCall-128         24.5µs ± 2%    23.6µs ± 2%   -3.68%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128       32.4µs ± 1%    31.6µs ± 2%   -2.44%  (p=0.016 n=5+5)
Workloads/NoCalls-128            35.1µs ± 2%    34.0µs ± 2%   -3.29%  (p=0.008 n=5+5)
Workloads/OneCall-128            49.1µs ± 1%    47.3µs ± 3%   -3.71%  (p=0.016 n=5+5)

name                           old alloc/op   new alloc/op   delta
NewWorkload/NoCallsNoGen-128      0.00B          0.00B          ~     (all equal)
NewWorkload/NoCalls-128            346B ± 0%      264B ± 0%  -23.70%  (p=0.008 n=5+5)
NewWorkload/OneCall-128            362B ± 0%      280B ± 0%  -22.65%  (p=0.008 n=5+5)
NewSimulator/NoCallsNoGen-128    6.00kB ± 0%    6.00kB ± 0%     ~     (all equal)
NewSimulator/NoCalls-128         6.39kB ± 0%    6.26kB ± 0%   -1.99%  (p=0.008 n=5+5)
NewSimulator/OneCall-128         6.46kB ± 0%    6.33kB ± 0%     ~     (p=0.079 n=4+5)
Workloads/NoCallsNoGen-128       6.52kB ± 0%    6.52kB ± 0%     ~     (all equal)
Workloads/NoCalls-128            7.02kB ± 0%    6.93kB ± 0%   -1.17%  (p=0.008 n=5+5)
Workloads/OneCall-128            8.36kB ± 0%    8.28kB ± 0%   -0.97%  (p=0.008 n=5+5)

name                           old allocs/op  new allocs/op  delta
NewWorkload/NoCallsNoGen-128       0.00           0.00          ~     (all equal)
NewWorkload/NoCalls-128            9.00 ± 0%      6.00 ± 0%  -33.33%  (p=0.008 n=5+5)
NewWorkload/OneCall-128            10.0 ± 0%       7.0 ± 0%  -30.00%  (p=0.008 n=5+5)
NewSimulator/NoCallsNoGen-128      21.0 ± 0%      21.0 ± 0%     ~     (all equal)
NewSimulator/NoCalls-128           30.0 ± 0%      27.0 ± 0%  -10.00%  (p=0.008 n=5+5)
NewSimulator/OneCall-128           32.0 ± 0%      29.0 ± 0%   -9.38%  (p=0.008 n=5+5)
Workloads/NoCallsNoGen-128         36.0 ± 0%      36.0 ± 0%     ~     (all equal)
Workloads/NoCalls-128              49.0 ± 0%      46.0 ± 0%   -6.12%  (p=0.008 n=5+5)
Workloads/OneCall-128              83.0 ± 0%      80.0 ± 0%   -3.61%  (p=0.008 n=5+5)
```